개시글 검색 추가

+++ osoriweb/website/board/views.py
@@ -1,7 +1,7 @@
 from django.shortcuts import render, get_object_or_404, redirect
 from django.utils import timezone
 from .models import InfoArticle, InfoComment, CommentLikes
-from .forms import InfoForm, InfoCommentForm
+from .forms import InfoForm, InfoCommentForm, SearchForm
 from django.contrib.auth.decorators import login_required
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 
@@ -13,19 +13,28 @@ def noti_board(request):
     return render(request, 'board/noti_board.html', {})
 
 def info_board(request):
-	posts = InfoArticle.objects.filter(created_date__lte=timezone.now()).order_by('-created_date')
-	paginator = Paginator(posts, 25) # Show 25 articles per page
+    if request.method == "POST":
+        form = SearchForm(request.POST)
+        if form.is_valid():
+            arg = form.cleaned_data
+            key = arg.get('searchKey')
+            posts = InfoArticle.objects.filter(created_date__lte = timezone.now()).filter(title__icontains = key).order_by('-created_date')
+    else:
+        posts = InfoArticle.objects.filter(created_date__lte=timezone.now()).order_by('-created_date')
+        form = SearchForm()
 
-	page = request.GET.get('page')
-	try:
-	    articles = paginator.page(page)
-	except PageNotAnInteger:
-	    # If page is not an integer, deliver first page.
-	    articles = paginator.page(1)
-	except EmptyPage:
-	    # If page is out of range (e.g. 9999), deliver last page of results.
-	    articles = paginator.page(paginator.num_pages)
-	return render(request, 'board/info_board.html', {'articles' : articles})
+    paginator = Paginator(posts, 25) # Show 25 articles per page
+
+    page = request.GET.get('page')
+    try:
+        articles = paginator.page(page)
+    except PageNotAnInteger:
+        # If page is not an integer, deliver first page.
+        articles = paginator.page(1)
+    except EmptyPage:
+        # If page is out of range (e.g. 9999), deliver last page of results.
+        articles = paginator.page(paginator.num_pages)
+    return render(request, 'board/info_board.html', {'articles' : articles, 'form' : form})
 
 @login_required
 def info_new(request):
@@ -89,14 +98,16 @@ def info_detail(request, pk):
 @login_required
 def info_remove(request, pk):
     article = get_object_or_404(InfoArticle, pk=pk)
-    article.delete()
+    if request.user == article.author or request.user.is_superuser:
+        article.delete()
     return redirect('info_board')
 
 @login_required
 def info_comment_remove(request, pk):
     comment = get_object_or_404(InfoComment, pk = pk)
     article = comment.article
-    comment.delete()
+    if request.user == comment.author or request.user.is_superuser:
+        comment.delete()
     return redirect('info_detail', pk=article.pk)
 
 @login_required
